move animation thread to top level conftest
